Rename Transactions to TransactionsPage
